+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ffc-sfi-payments",
-  "version": "2.6.4",
+  "version": "2.6.5",
   "description": "FFC SFI payment services",
   "homepage": "https://github.com/DEFRA/ffc-sfi-payments",
   "main": "app/index.js",
